Added Backend and BackendInterface files
BackendInterface is slightly changed from original proposal due to foolish type mistakes.  Backend relies on slightly modified BuildingInterface object (getTypes returning a single String instead of an array of Strings), and includes a few methods not specified in the original proposal.  A pair of methods for getting the list of buildings along a path (one between specific points, the other to a general type) was added, as was a passthrough method for getBuildingSpecifiers, which is presumably used in the Frontend.