Added DW Interfaces
Putting these in largely for reference (including to my change to BuildingInterface) and because I'd written them out.  The getTypes() method in BuildingInterface was changed to return a single String rather than an array of Strings, as returning an array of single-character Strings leads to added code complexity for no gain.